use test mixin for test cases that mock uses of requests library
