+++ apps/member/tests/test_views.py
@@ -1,18 +1,15 @@
-from httmock import all_requests, HTTMock
+from httmock import HTTMock
 
 from django.test import TestCase
 from django.urls.exceptions import NoReverseMatch
 from django.urls import reverse
 
-from apps.common.tests.base import SMHAppTestMixin
+from apps.common.tests.base import MockResourceDataMixin, SMHAppTestMixin
 from apps.common.tests.factories import UserFactory
 from apps.org.models import (
     ResourceGrant, REQUEST_APPROVED, REQUEST_DENIED, REQUEST_REQUESTED
 )
-from apps.org.tests.factories import (
-    OrganizationFactory, ResourceGrantFactory, ResourceRequestFactory,
-    UserSocialAuthFactory
-)
+from apps.org.tests.factories import ResourceGrantFactory, ResourceRequestFactory
 from apps.sharemyhealth.resources import Resource
 
 
@@ -253,55 +250,8 @@ class RevokeResourceRequestTestCase(SMHAppTestMixin, TestCase):
         self.assertIsNone(getattr(self.resource_request, 'resourcegrant', None))
 
 
-class RecordsViewTestCase(SMHAppTestMixin, TestCase):
+class RecordsViewTestCase(MockResourceDataMixin, SMHAppTestMixin, TestCase):
     url_name = 'member:records'
-
-    def setUp(self):
-        super().setUp()
-        # When we mock uses of the requests library in this test class, this is
-        # the expected mock response.
-        self.expected_response_success = {
-            'status_code': 200,
-            'content': {'test_key': 'Test content'}
-        }
-
-    @all_requests
-    def response_content_success(self, url, request):
-        return self.expected_response_success
-
-    def give_self_user_access_to_member_token(self, member, provider_name):
-        """
-        Give the self.user access to the member's access_token, and return access_token.
-
-        This method creates the necessary database objects so that the self.user
-        is in an Organization that has a ResourceGrant for the member's Resource,
-        so the self.user can use the member's access_token to get data about the
-        member.
-        """
-        # The self.user is a part of an Organization
-        organization = OrganizationFactory()
-        organization.users.add(self.user)
-        # The member has received an access_token to get their own data.
-        provider_name = provider_name
-        access_token = 'accessTOKENhere'
-        UserSocialAuthFactory(
-            user=member,
-            provider=provider_name,
-            extra_data={'refresh_token': 'refreshTOKEN', 'access_token': access_token}
-        )
-        # The member has approved the Organization's request for the member's data
-        resource_request = ResourceRequestFactory(
-            member=member,
-            organization=organization,
-            resourcegrant=None,
-            status=REQUEST_APPROVED
-        )
-        ResourceGrantFactory(
-            member=resource_request.member,
-            organization=resource_request.organization,
-            resource_class_path=resource_request.resource_class_path,
-            resource_request=resource_request
-        )
 
     def test_context_data(self):
         """GETting records view puts response of get_member_data() into the context."""
@@ -309,7 +259,7 @@ class RecordsViewTestCase(SMHAppTestMixin, TestCase):
         member = UserFactory()
         # Give the self.user access to the member's access_token.
         provider_name = Resource.name
-        self.give_self_user_access_to_member_token(member, provider_name)
+        self.give_user_access_to_member_token(self.user, member, provider_name)
 
         url = reverse(self.url_name, kwargs={'pk': member.pk})
         # We mock the use of the requests library, so we don't make real
@@ -331,7 +281,7 @@ class RecordsViewTestCase(SMHAppTestMixin, TestCase):
         member = UserFactory()
         # Give the self.user access to the member's access_token.
         provider_name = Resource.name
-        self.give_self_user_access_to_member_token(member, provider_name)
+        self.give_user_access_to_member_token(self.user, member, provider_name)
         url = reverse(self.url_name, kwargs={'pk': member.pk})
 
         with self.subTest('Authenticated'):
@@ -349,55 +299,8 @@ class RecordsViewTestCase(SMHAppTestMixin, TestCase):
             self.assertRedirects(response, expected_redirect)
 
 
-class DataSourcesViewTestCase(SMHAppTestMixin, TestCase):
+class DataSourcesViewTestCase(MockResourceDataMixin, SMHAppTestMixin, TestCase):
     url_name = 'member:data-sources'
-
-    def setUp(self):
-        super().setUp()
-        # When we mock uses of the requests library in this test class, this is
-        # the expected mock response.
-        self.expected_response_success = {
-            'status_code': 200,
-            'content': {'test_key': 'Test content'}
-        }
-
-    @all_requests
-    def response_content_success(self, url, request):
-        return self.expected_response_success
-
-    def give_self_user_access_to_member_token(self, member, provider_name):
-        """
-        Give the self.user access to the member's access_token, and return access_token.
-
-        This method creates the necessary database objects so that the self.user
-        is in an Organization that has a ResourceGrant for the member's Resource,
-        so the self.user can use the member's access_token to get data about the
-        member.
-        """
-        # The self.user is a part of an Organization
-        organization = OrganizationFactory()
-        organization.users.add(self.user)
-        # The member has received an access_token to get their own data.
-        provider_name = provider_name
-        access_token = 'accessTOKENhere'
-        UserSocialAuthFactory(
-            user=member,
-            provider=provider_name,
-            extra_data={'refresh_token': 'refreshTOKEN', 'access_token': access_token}
-        )
-        # The member has approved the Organization's request for the member's data
-        resource_request = ResourceRequestFactory(
-            member=member,
-            organization=organization,
-            resourcegrant=None,
-            status=REQUEST_APPROVED
-        )
-        ResourceGrantFactory(
-            member=resource_request.member,
-            organization=resource_request.organization,
-            resource_class_path=resource_request.resource_class_path,
-            resource_request=resource_request
-        )
 
     def test_get_parameters(self):
         """GETting data sources view with/without resource_name and record_type parameters."""
@@ -405,7 +308,7 @@ class DataSourcesViewTestCase(SMHAppTestMixin, TestCase):
         member = UserFactory()
         # Give the self.user access to the member's access_token.
         provider_name = Resource.name
-        self.give_self_user_access_to_member_token(member, provider_name)
+        self.give_user_access_to_member_token(self.user, member, provider_name)
 
         with self.subTest('no resource_name or record_type parameters'):
             url = reverse(self.url_name, kwargs={'pk': member.pk})
@@ -477,7 +380,7 @@ class DataSourcesViewTestCase(SMHAppTestMixin, TestCase):
         member = UserFactory()
         # Give the self.user access to the member's access_token.
         provider_name = Resource.name
-        self.give_self_user_access_to_member_token(member, provider_name)
+        self.give_user_access_to_member_token(self.user, member, provider_name)
         url = reverse(
             self.url_name,
             kwargs={
